Add pointer event regions to EventRegion https://bugs.webkit.org/show_bug.cgi?id=288616 rdar://145671088
Reviewed by Simon Fraser.

Add pointer event and synthetic mouse regions to Event region and populate it per layer based
on the RenderStyle value of eventListenerRegionTypes. Refactor EventRegion to hold a
EventTrackingRegions for touch event listeners.

* Source/WTF/wtf/PlatformEnableCocoa.h:
* Source/WebCore/page/scrolling/ScrollingTree.h:
* Source/WebCore/platform/EventTrackingRegions.cpp:
(WebCore::EventTrackingRegions::uniteEventListeners):
(WebCore::operator<<):
* Source/WebCore/platform/EventTrackingRegions.h:
* Source/WebCore/rendering/EventRegion.cpp:
(WebCore::EventRegion::EventRegion):
(WebCore::containsTouchEventType):
(WebCore::EventRegion::uniteEventListeners):
(WebCore::EventRegion::eventListenerRegionTypesForPoint const): Deleted.
(WebCore::EventRegion::eventListenerRegionForType const): Deleted.
(WebCore::EventRegion::containsEditableElementsInRect const): Deleted.
(WebCore::EventRegion::appendInteractionRegions): Deleted.
(WebCore::EventRegion::clearInteractionRegions): Deleted.
(WebCore::EventRegion::dump const): Deleted.
(WebCore::operator<<): Deleted.
* Source/WebCore/rendering/EventRegion.h:
* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::patternForEventListenerRegionType):
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/RenderStyleConstants.cpp:
(WebCore::operator<<):
* Source/WebCore/rendering/style/RenderStyleConstants.h:
* Source/WebCore/style/StyleAdjuster.cpp:
(WebCore::Style::Adjuster::computeEventListenerRegionTypes):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/UIProcess/RemoteLayerTree/ios/RemoteLayerTreeViews.h:

Canonical link: https://commits.webkit.org/291377@main

+++ Source/WebKit/UIProcess/RemoteLayerTree/cocoa/RemoteScrollingTreeCocoa.h
@@ -29,7 +29,7 @@
 
 namespace WebCore {
 class EventRegion;
-enum class EventListenerRegionType : uint16_t;
+enum class EventListenerRegionType : uint32_t;
 class FloatPoint;
 }
